krableadsV2/requirements.txt
python-telegram-bot[job-queue]>=20.0
supabase>=2.28.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2025.1
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
pymupdf>=1.24.0

